Similar future fix to #173 for other avail actions
The standard Clusterware repos for each of these things happen to not
yet have contained anything other than the subdirectories containing
their content, and so a similar bug to #173 has not yet arisen for
these. However, other files could easily also be added to these, or
users could use their own repos containing other files, and then this
bug would reoccur. This commit pre-emptively adds the same fix to stop
this from happening.

Potentially we could write these in a way so that we don't need to apply
changes such as this in multiple locations, but I can't think of the
best way to do this right now.