Fix issue where screenshot over multiple lines
Previously, the pipeline which produces session screenshots was giving
an output of multiple base64-encoded lines which together encode the
screenshot. These were then being passed into running the
session-screenshot handler hooks. In particular, this then meant the
flight/session-screenshot handler would run for each line in turn
independently, causing ~15 requests to be POSTed to the given URL, each
including a single line of the base64-encoded screenshot.

This commit changes this so that these lines are joined into a single
line before being passed to the session-screenshot handler hook, and so
only a single request will be made. Internal whitespace is also stripped
from this string as it is unnecessary.